// src/types/napcat.ts
export interface OB11BaseEvent {
  time: number;
  self_id: OneBotId;
  post_type: string;
}

export interface OB11Message extends OB11BaseEvent {
  post_type: 'message' | 'message_sent';
  message_type: 'private' | 'group';
  message_id: OneBotMessageId;
  user_id: OneBotId;
  message: OB11MessageSegment[];
  raw_message: string;
  sender: OB11Sender;
  group_id?: OneBotId;
}

export interface OB11Sender {
  user_id: OneBotId;
  nickname: string;
  card?: string;
  role?: 'owner' | 'admin' | 'member';
}

export interface OB11MessageSegment {
  type: string;
  data: Record<string, unknown>;
}

export type OB11Event = OB11RequestEvent | OB11NoticeEvent;

export interface OB11RequestEvent extends OB11BaseEvent {
  post_type: 'request';
  request_type: 'group' | 'friend';
  sub_type?: string;
  group_id?: OneBotId;
  user_id: OneBotId;
  comment: string;
  flag: string;
}

export interface OB11NoticeEvent extends OB11BaseEvent {
  post_type: 'notice';
  notice_type: string;
  group_id?: OneBotId;
  user_id?: OneBotId;
  operator_id?: OneBotId;
  sub_type?: string;
}